job_manager_proj/pyproject.toml
[tool.poetry]
name = "job-manager"
version = "0.1.0"
description = ""
authors = ["Alexandr <maximchik.alexandr@yandex.ru>"]
readme = "README.md"
packages = [{include = "job_manager"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
django-import-export = "^3.2.0"
docxtpl = "^0.16.7"
celery = "^5.3.1"
redis = "^4.6.0"
django-celery-beat = "^2.5.0"
yadisk = "^1.3.3"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"
ipython = "^8.12.0"
pytest-django = "^4.5.2"
prospector = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "job_manager_proj.settings"
python_files = [
    "tests.py",
    "test*.py",
    "*_tests.py"
]
addopts = [
    "--nomigrations",
    "--create-db",
    "-p no:warnings"
]